LLM/requirements.txt
llama-cpp-python
huggingface-hub
sentence-transformers
faiss-cpu
pypdf
numpy
